correct use of ptrdiff_t for nrrdPad
git-svn-id: https://svn.code.sf.net/p/teem/code/teem/trunk@2743 3d70eeeb-363e-0410-a505-8a46323a89f2

+++ src/ten/chan.c
@@ -277,7 +277,7 @@ tenEMatrixCalc(Nrrd *nemat, const Nrrd *_nbmat, int knownB0) {
   char me[]="tenEMatrixCalc", err[AIR_STRLEN_MED];
   Nrrd *nbmat, *ntmp;
   airArray *mop;
-  int padmin[2], padmax[2];
+  ptrdiff_t padmin[2], padmax[2];
   unsigned int ri;
   double *bmat;
